backend/worker/requirements.txt
fastapi
requests
celery[redis]
docxtpl
pillow
dropbox
